Bug in userService as saving the entry for a user is updating the password so need to create new method while saving user to insert journal entries.
**Adding Authentication:**
Only changed JournalEntryController_V2 removed username which is taken from authorization and made sure the user can only access his records only not others.

+++ src/main/java/com/example/journalApp/service/JournalEntryService.java
@@ -45,10 +45,14 @@ public class JournalEntryService {
         return journalEntryRepository.findById(id);
     }
 
-    public void deleteById(ObjectId id, String username){
+    public Boolean deleteById(ObjectId id, String username){
+        boolean removed = false;
         User user = userService.findByUsername(username);
-        user.getJournalEntries().removeIf(x -> x.getId().equals(id));
-        userService.saveEntry(user);
-        journalEntryRepository.deleteById(id);
+        removed = user.getJournalEntries().removeIf(x -> x.getId().equals(id));
+        if(removed){
+            userService.saveEntry(user);
+            journalEntryRepository.deleteById(id);
+        }
+        return removed;
     }
 }
